requirements.txt
# pip install -r requirements.txt pour tout installer directement


discord.py
python-dotenv
